Tests: Move BoardStateTests into more appropriate places
Originally, BoardStateTests tested functionality that existed in
BoardState, but now that functionality lives in other places. This
commit moves tests to corresponding places.
