+++ Assets/Scripts/StandaloneRules/StandaloneRulesTests/TestUtils.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 865b5585aa4cf4bfab8f7af88653a0fa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
